Refactor as per PR comments
