+++ kroxylicious-runtime/src/test/java/io/kroxylicious/proxy/internal/reload/AddClusterTest.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.kroxylicious.proxy.internal.reload.operations;
+package io.kroxylicious.proxy.internal.reload;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
